\name{QuasR-package}
\alias{QuasR-package}
\alias{QuasR}
\docType{package}

\title{Quantify and Annotate Short Reads in R}

\description{
This package provides a pipeline for the quantification and analysis of Short Reads.
}

\details{
See packageDescription('QuasR') for package details.
}

\author{
Anita Lerch, Dimos Gaidatzis and Michael Stadler
}

%\references{TODO}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
  \code{\link{qAlign}}, \code{\link{qCount}}, \code{\link{qProfile}}, \code{\link{qMeth}}, \code{\link{qQCReport}}
}

\examples{
\dontrun{
    # see qCount, qMeth and qProfile manual pages for examples
    example(qCount)
    example(qMeth)
    example(qProfile)
}
}
